csp: add csp_send_free() function
The csp_send() function will consume and subsequently free the buffer
element if transmission is succesful. If the transmission fails however,
it is the users responsibility to either free or retransmit the buffer.
The typical case is to just free the element on error, which leads to
lots of code like this:

   if (!csp_send(conn, packet, timeout))
       csp_buffer_free(packet);

This patch adds a new function, csp_send_free(), with the same arguments
as csp_send(), that always frees the buffer element.

Signed-off-by: Jeppe Ledet-Pedersen <jlp@satlab.com>